pyproject.toml
[tool.poetry]
name = "escrow"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.104.1"
uvicorn = {extras = ["standard"], version = "0.24.0"}
python-multipart = "0.0.6"
pyjwt = "2.8.0"
python-dotenv = "1.0.0"
pydantic-settings = "2.1.0"
jinja2 = "3.1.2"
typing-extensions = ">=4.0.0"
sqlalchemy = "2.0.23"
alembic = "1.12.1"
asyncpg = "0.29.0"
psycopg2-binary = "2.9.9"
cryptography = "41.0.7"
redis = "5.0.1"
tronpy = ">=0.4.0"
bcrypt = "4.0.1"
passlib = "1.7.4"
Pillow = ">=10.0.0"
ecdsa = ">=0.18.0"
base58 = ">=2.1.1"
pycryptodome = ">=3.19.0"
aiohttp = ">=3.9.0"
bip32 = ">=2.0.0"
mnemonic = ">=0.20"
didcomm = {path = "didcomm", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "network: тесты с реальными HTTP-запросами (pytest -m network)",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
